pyproject.toml
[tool.poetry]
name = "python-kafka"
version = "0.1.0"
description = ""
authors = ["Itay Ozery <prowler421@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
kafka-python = "^2.0.2"
confluent-kafka = "^1.9.0"
fastapi = "^0.79.0"
uvicorn = "^0.18.2"
requests = "^2.28.1"
jsonschema = "^4.7.2"
python-dotenv = "^0.20.0"
asyncio = "^3.4.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
